cleaned test for course
